jobs/requirements.txt
aiodns==2.0.0
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
ccxt==1.39.54
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
idna==2.10
install==1.3.4
multidict==5.1.0
psycopg2==2.8.6
pycares==3.1.1
pycodestyle==2.6.0
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.1.0
